requirements.txt
alembic==1.15.2
bandit==1.8.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
flake8==7.2.0
flake8-json==24.4.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
mando==0.7.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
packaging==24.2
pbr==6.1.1
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
radon==6.0.1
requests==2.32.3
rich==14.0.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.40
stevedore==5.4.1
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
